feat(desktop): fantasy in-widget config — one bar + Yahoo wizard in-feed [REL-37]
- Scroll fix: root h-full → min-h-full; the inner flex-1 overflow-y-auto
  content scrollport is gone — PageLayout owns scrolling and the bar
  actually pins (sentinel elevation).
- ONE WidgetBar (the secondary LeagueSwitcher pill band is deleted):
  Segmented [Overview | Matchup | Standings | Roster] · league
  SelectMenu (emoji + name, ★ primary, live dot in rows; hidden on
  Overview / single league / account view) · "N live" pulse · gear.
- Gear popover: primary league, enabled leagues (empty = all,
  round-trips through the same [] convention as ConnectedView), default
  view, and a "Yahoo account & leagues…" entry row that opens the
  wizard in-feed (GearMenu now supports function-children to close the
  popover on navigation; SelectMenu labels widened to ReactNode).
- YahooConnectFlow.tsx: the ConfigPanel's ENTIRE OAuth state machine
  (disconnected → discovering → picking → importing → connected) moved
  verbatim, de-tiered (caps Infinity since 2026-07-02 — maxFantasy /
  remainingCapacity / atLeagueLimit / free-tier gate deleted; slices
  and toggle guards dropped). LeaguePicker + ConnectedView de-tiered to
  match. Mounted in-feed: it IS the feed when nothing is imported
  ("Connect Yahoo" inline, no navigation), and the account view from
  the gear. ConfigPanel is now a thin wrapper around it (parity until
  teardown).
- "No leagues enabled" dead-end now self-heals: the CTA enables all
  leagues directly instead of pointing at Configure.

Gate: tsc clean, vitest 490/490, all five Playwright suites green.
Pending before merge: manual smoke of the wizard's 5 phases against the
live dev stack (state machine moved verbatim — per plan, no harness).
