+++ desktop/src/channels/fantasy/FeedTab.tsx
@@ -7,16 +7,18 @@
  *   - Standings: playoff-aware standings for the primary league
  *   - Roster:   user's (or any team's) roster with injury spotlight
  *
- * A league switcher appears when the user has 2+ active leagues, so each
- * sub-view follows a single "current" league.  The Feed is deliberately
- * read-only; import / disconnect / tier gating lives on the Configure
- * tab.
+ * ONE Kalshi-style control bar (widget-bar primitives): Segmented
+ * sub-tabs · league SelectMenu (emoji + name, ★ primary, live dot;
+ * hidden on Overview / single league) · "N live" pulse · gear popover
+ * (primary league, enabled leagues, default view, and the "Yahoo
+ * account & leagues…" entry). The Yahoo OAuth/import wizard
+ * (YahooConnectFlow) mounts IN-FEED — as the whole feed when nothing
+ * is connected, or as the account view from the gear.
  */
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { clsx } from "clsx";
 import {
   Activity,
-  ClipboardList,
   LayoutGrid,
   Star,
   Swords,
@@ -28,10 +30,19 @@ import { useQuery } from "@tanstack/react-query";
 import { dashboardQueryOptions } from "../../api/queries";
 import { useShell } from "../../shell-context";
 import EmptyChannelState from "../../components/EmptyChannelState";
+import { WidgetBar } from "../../components/widget-bar/Bar";
+import { MenuHeading, MenuRow } from "../../components/widget-bar/Menu";
+import {
+  Segmented,
+  type SegmentedOption,
+} from "../../components/widget-bar/Segmented";
+import { SelectMenu } from "../../components/widget-bar/SelectMenu";
+import { GearMenu } from "../../components/widget-bar/GearMenu";
 import { OverviewView } from "./OverviewView";
 import { MatchupView } from "./MatchupView";
 import { StandingsView } from "./StandingsView";
 import { RosterView } from "./RosterView";
+import YahooConnectFlow from "./YahooConnectFlow";
 import {
   SPORT_EMOJI,
   isMatchupLive,
@@ -39,8 +50,11 @@ import {
 } from "./types";
 import { filterEnabledLeagues, resolvePrimaryLeague } from "./view";
 import type { FeedTabProps, ChannelManifest } from "../../types";
-import type { FantasySubTab } from "../../preferences";
+import type { FantasyDisplayPrefs, FantasySubTab } from "../../preferences";
 import type { LeagueResponse, MyLeaguesResponse } from "./types";
+
+/** Channel accent — kept in sync with `fantasyChannel.hex`. */
+const FANTASY_HEX = "#6366f1";
 
 // ── Channel manifest ─────────────────────────────────────────────
 
@@ -76,13 +90,7 @@ function extractLeagues(data: unknown): LeagueResponse[] {
   return [];
 }
 
-interface SubTabMeta {
-  value: FantasySubTab;
-  label: string;
-  icon: typeof LayoutGrid;
-}
-
-const SUB_TABS: SubTabMeta[] = [
+const SUB_TABS: SegmentedOption<FantasySubTab>[] = [
   { value: "overview", label: "Overview", icon: LayoutGrid },
   { value: "matchup", label: "Matchup", icon: Swords },
   { value: "standings", label: "Standings", icon: Trophy },
@@ -91,9 +99,10 @@ const SUB_TABS: SubTabMeta[] = [
 
 // ── FeedTab ──────────────────────────────────────────────────────
 
-function FantasyFeedTab({ mode, feedContext, onConfigure }: FeedTabProps) {
-  const { prefs } = useShell();
+function FantasyFeedTab({ mode, feedContext }: FeedTabProps) {
+  const { prefs, onPrefsChange } = useShell();
   const dp = prefs.channelDisplay.fantasy;
+  const isComfort = mode === "comfort";
 
   const { data: dashboard } = useQuery(dashboardQueryOptions());
   const fantasyData = dashboard?.data?.fantasy;
@@ -147,105 +156,138 @@ function FantasyFeedTab({ mode, feedContext, onConfigure }: FeedTabProps) {
     setSubTab("matchup");
   }, []);
 
-  // ── Empty / loading states ───────────────────────────────────
-  if (leagues.length === 0) {
-    return (
-      <EmptyChannelState
-        refreshing={Boolean(feedContext.__refreshing)}
-        icon={Swords}
-        noun="fantasy leagues"
-        hasConfig={!!feedContext.__hasConfig}
-        dashboardLoaded={!!feedContext.__dashboardLoaded}
-        loadingNoun="leagues"
-        actionHint="connect your Yahoo account"
-        onConfigure={onConfigure}
-      />
-    );
-  }
+  // Account view (the Yahoo OAuth/import wizard) — opened from the gear.
+  const [accountOpen, setAccountOpen] = useState(false);
+  const pickSubTab = useCallback((t: FantasySubTab) => {
+    setSubTab(t);
+    setAccountOpen(false);
+  }, []);
 
-  if (visibleLeagues.length === 0) {
-    return (
-      <div className="flex h-full flex-col items-center justify-center gap-3 p-6 text-center">
-        <Activity size={28} className="text-fg-3" />
-        <div className="text-[13px] font-semibold text-fg">
-          No leagues enabled for viewing
-        </div>
-        <p className="max-w-sm text-[11px] text-fg-3">
-          Every one of your imported leagues is currently hidden. Enable at
-          least one in the Configure tab to see your matchups, standings, and
-          rosters here.
-        </p>
-        <button
-          type="button"
-          onClick={onConfigure}
-          className="mt-1 rounded-md bg-accent/10 px-3 py-1.5 text-[11px] font-medium text-accent hover:bg-accent/20 cursor-pointer"
-        >
-          Open Configure
-        </button>
-      </div>
-    );
-  }
+  const enableAllLeagues = useCallback(() => {
+    onPrefsChange({
+      ...prefs,
+      channelDisplay: {
+        ...prefs.channelDisplay,
+        fantasy: { ...prefs.channelDisplay.fantasy, enabledLeagueKeys: [] },
+      },
+    });
+  }, [prefs, onPrefsChange]);
+
+  // League options for the bar's SelectMenu — emoji + name, ★ on the
+  // primary, live dot on leagues with a live matchup.
+  const leagueOptions = useMemo(
+    () =>
+      visibleLeagues.map((l) => {
+        const ctx = userMatchupContext(l);
+        const live = ctx && isMatchupLive(ctx.matchup);
+        const isPrimary = l.league_key === primaryLeague?.league_key;
+        return {
+          value: l.league_key,
+          label: (
+            <span className="inline-flex min-w-0 items-center gap-1.5">
+              <span aria-hidden>{SPORT_EMOJI[l.game_code] ?? "🏆"}</span>
+              <span className="truncate">{l.name}</span>
+              {isPrimary && (
+                <Star size={10} className="shrink-0 fill-accent stroke-accent" />
+              )}
+              {live && (
+                <span className="h-1.5 w-1.5 shrink-0 animate-pulse rounded-full bg-live" />
+              )}
+            </span>
+          ),
+        };
+      }),
+    [visibleLeagues, primaryLeague?.league_key],
+  );
 
   const liveCount = visibleLeagues.filter((l) => {
     const ctx = userMatchupContext(l);
     return ctx && isMatchupLive(ctx.matchup);
   }).length;
 
+  const dataPending =
+    !feedContext.__dashboardLoaded || Boolean(feedContext.__refreshing);
+  // Nothing imported: the wizard IS the feed ("Connect Yahoo" inline).
+  const showConnect = leagues.length === 0 && !dataPending;
+  const showLoading = leagues.length === 0 && dataPending;
+  const showLeagueSelect =
+    !accountOpen &&
+    !showConnect &&
+    subTab !== "overview" &&
+    visibleLeagues.length > 1 &&
+    activeLeague !== null;
+
   return (
-    <div className={clsx("flex h-full flex-col", mode === "compact" && "text-[12px]")}>
-      {/* Top bar: sub-tabs + live pulse */}
-      <div className="sticky top-0 z-20 flex items-center gap-2 border-b border-edge/30 bg-surface px-3 py-2">
-        <div className="flex gap-1">
-          {SUB_TABS.map((tab) => {
-            const Icon = tab.icon;
-            const disabled = tab.value !== "overview" && !activeLeague;
-            const active = subTab === tab.value;
-            return (
-              <button
-                key={tab.value}
-                type="button"
-                onClick={() => setSubTab(tab.value)}
-                disabled={disabled}
-                className={clsx(
-                  "flex items-center gap-1.5 rounded-lg px-3 py-1.5 text-[12px] font-medium transition-colors",
-                  active
-                    ? "bg-accent/15 text-accent"
-                    : "text-fg-3 hover:bg-surface-hover hover:text-fg-2",
-                  disabled && "cursor-not-allowed opacity-50",
-                )}
-              >
-                <Icon size={13} />
-                {tab.label}
-              </button>
-            );
-          })}
-        </div>
+    // NO inner scroll container: the Source page (PageLayout) owns the
+    // scroll — sticky pins against it.
+    <div className={clsx("flex min-h-full flex-col", mode === "compact" && "text-[12px]")}>
+      {isComfort && (
+        <WidgetBar>
+          <Segmented
+            ariaLabel="Fantasy view"
+            value={subTab}
+            onChange={pickSubTab}
+            options={SUB_TABS}
+          />
 
-        <div className="ml-auto flex items-center gap-3 font-mono text-[10px] uppercase tracking-wider text-fg-3">
-          <span>
-            {visibleLeagues.length} league{visibleLeagues.length === 1 ? "" : "s"}
-          </span>
-          {liveCount > 0 && (
-            <span className="inline-flex items-center gap-1 text-live">
-              <span className="h-1.5 w-1.5 rounded-full bg-live animate-pulse" />
-              {liveCount} live
-            </span>
+          {showLeagueSelect && activeLeague && (
+            <SelectMenu
+              value={activeLeague.league_key}
+              options={leagueOptions}
+              onChange={setActiveLeagueKey}
+              ariaLabel="Switch league"
+            />
           )}
-        </div>
-      </div>
 
-      {/* Secondary bar: league switcher when 2+ leagues AND not in overview */}
-      {visibleLeagues.length > 1 && subTab !== "overview" && (
-        <LeagueSwitcher
-          leagues={visibleLeagues}
-          activeKey={activeLeague?.league_key ?? null}
-          primaryKey={primaryLeague?.league_key ?? null}
-          onSelect={setActiveLeagueKey}
-        />
+          <div className="ml-auto flex min-w-0 shrink items-center gap-3">
+            {liveCount > 0 && (
+              <span className="inline-flex items-center gap-1 font-mono text-[10px] uppercase tracking-wider text-live">
+                <span className="h-1.5 w-1.5 rounded-full bg-live animate-pulse" />
+                {liveCount} live
+              </span>
+            )}
+            <FantasyGear
+              leagues={leagues}
+              onOpenAccount={() => setAccountOpen(true)}
+            />
+          </div>
+        </WidgetBar>
       )}
 
-      {/* Content */}
-      <div className="flex-1 overflow-y-auto">
+      {showLoading ? (
+        <div className="flex flex-1 flex-col justify-center">
+          <EmptyChannelState
+            refreshing={Boolean(feedContext.__refreshing)}
+            icon={Swords}
+            noun="fantasy leagues"
+            hasConfig={!!feedContext.__hasConfig}
+            dashboardLoaded={!!feedContext.__dashboardLoaded}
+            loadingNoun="leagues"
+          />
+        </div>
+      ) : showConnect || accountOpen ? (
+        <div className="pt-4">
+          <YahooConnectFlow hex={FANTASY_HEX} />
+        </div>
+      ) : visibleLeagues.length === 0 ? (
+        <div className="flex flex-1 flex-col items-center justify-center gap-3 p-6 text-center">
+          <Activity size={28} className="text-fg-3" />
+          <div className="text-[13px] font-semibold text-fg">
+            No leagues enabled for viewing
+          </div>
+          <p className="max-w-sm text-[11px] text-fg-3">
+            Every one of your imported leagues is currently hidden. Enable
+            them in the gear menu, or show them all:
+          </p>
+          <button
+            type="button"
+            onClick={enableAllLeagues}
+            className="mt-1 rounded-md bg-accent/10 px-3 py-1.5 text-[11px] font-medium text-accent hover:bg-accent/20 cursor-pointer"
+          >
+            Show all leagues
+          </button>
+        </div>
+      ) : (
         <motion.div
           key={subTab + (activeLeague?.league_key ?? "none")}
           initial={{ opacity: 0, y: 4 }}
@@ -264,55 +306,124 @@ function FantasyFeedTab({ mode, feedContext, onConfigure }: FeedTabProps) {
           {subTab === "standings" && <StandingsView league={activeLeague} />}
           {subTab === "roster" && <RosterView league={activeLeague} />}
         </motion.div>
-      </div>
+      )}
     </div>
   );
 }
 
-// ── League switcher pill bar ─────────────────────────────────────
+// ── Gear popover (display prefs + account entry) ────────────────
+//
+// primaryLeagueKey / enabledLeagueKeys / defaultSubTab previously had
+// their only edit surface buried in the Configure page's ConnectedView
+// (which keeps them — same prefs, no conflict). The gear puts them one
+// click from the feed, plus the entry into the Yahoo account view.
 
-interface LeagueSwitcherProps {
+function FantasyGear({
+  leagues,
+  onOpenAccount,
+}: {
   leagues: LeagueResponse[];
-  activeKey: string | null;
-  primaryKey: string | null;
-  onSelect: (key: string) => void;
-}
+  onOpenAccount: () => void;
+}) {
+  const { prefs, onPrefsChange } = useShell();
+  const dp = prefs.channelDisplay.fantasy;
 
-function LeagueSwitcher({ leagues, activeKey, primaryKey, onSelect }: LeagueSwitcherProps) {
+  const patch = useCallback(
+    (partial: Partial<FantasyDisplayPrefs>) => {
+      onPrefsChange({
+        ...prefs,
+        channelDisplay: {
+          ...prefs.channelDisplay,
+          fantasy: { ...prefs.channelDisplay.fantasy, ...partial },
+        },
+      });
+    },
+    [prefs, onPrefsChange],
+  );
+
+  // Empty enabled list means "all" — reflect that as fully enabled.
+  const enabledSet = useMemo(() => {
+    if (!dp.enabledLeagueKeys || dp.enabledLeagueKeys.length === 0) {
+      return new Set(leagues.map((l) => l.league_key));
+    }
+    return new Set(dp.enabledLeagueKeys);
+  }, [dp.enabledLeagueKeys, leagues]);
+
+  const toggleEnabled = useCallback(
+    (key: string) => {
+      const next = new Set(enabledSet);
+      if (next.has(key)) next.delete(key);
+      else next.add(key);
+      // Every league enabled is stored as [] ("all").
+      patch({
+        enabledLeagueKeys:
+          next.size === leagues.length ? [] : Array.from(next),
+      });
+    },
+    [enabledSet, leagues.length, patch],
+  );
+
   return (
-    <div className="flex items-center gap-2 overflow-x-auto border-b border-edge/30 bg-surface-2/40 px-3 py-1.5">
-      <ClipboardList size={12} className="shrink-0 text-fg-3" />
-      <div className="flex items-center gap-1">
-        {leagues.map((l) => {
-          const active = l.league_key === activeKey;
-          const isPrimary = l.league_key === primaryKey;
-          const ctx = userMatchupContext(l);
-          const live = ctx && isMatchupLive(ctx.matchup);
-          return (
-            <button
-              key={l.league_key}
-              type="button"
-              onClick={() => onSelect(l.league_key)}
-              className={clsx(
-                "flex shrink-0 items-center gap-1.5 rounded-full border px-2.5 py-1 text-[11px] font-medium transition-colors cursor-pointer",
-                active
-                  ? "border-accent/60 bg-accent/15 text-accent"
-                  : "border-edge/40 text-fg-3 hover:border-accent/40 hover:text-fg-2",
-              )}
-            >
-              <span aria-hidden>{SPORT_EMOJI[l.game_code] ?? "🏆"}</span>
-              <span className="max-w-[140px] truncate">{l.name}</span>
-              {isPrimary && (
-                <Star size={10} className="fill-accent stroke-accent" />
-              )}
-              {live && (
-                <span className="ml-0.5 h-1.5 w-1.5 animate-pulse rounded-full bg-live" />
-              )}
-            </button>
-          );
-        })}
-      </div>
-    </div>
+    <GearMenu ariaLabel="Fantasy settings" panelClassName="right-0 w-80">
+      {(close) => (
+        <>
+          {leagues.length > 0 && (
+            <>
+              <MenuHeading>Primary league</MenuHeading>
+              <MenuRow
+                selected={!dp.primaryLeagueKey}
+                onClick={() => patch({ primaryLeagueKey: null })}
+                role="menuitemradio"
+              >
+                Auto (first active)
+              </MenuRow>
+              {leagues.map((l) => (
+                <MenuRow
+                  key={l.league_key}
+                  selected={dp.primaryLeagueKey === l.league_key}
+                  onClick={() => patch({ primaryLeagueKey: l.league_key })}
+                  role="menuitemradio"
+                >
+                  {SPORT_EMOJI[l.game_code] ?? "🏆"} {l.name}
+                </MenuRow>
+              ))}
+              <MenuHeading>Enabled leagues</MenuHeading>
+              {leagues.map((l) => (
+                <MenuRow
+                  key={l.league_key}
+                  selected={enabledSet.has(l.league_key)}
+                  onClick={() => toggleEnabled(l.league_key)}
+                  role="menuitemcheckbox"
+                >
+                  {SPORT_EMOJI[l.game_code] ?? "🏆"} {l.name}
+                </MenuRow>
+              ))}
+              <MenuHeading>Default view</MenuHeading>
+              {SUB_TABS.map((t) => (
+                <MenuRow
+                  key={t.value}
+                  selected={(dp.defaultSubTab ?? "overview") === t.value}
+                  onClick={() => patch({ defaultSubTab: t.value })}
+                  role="menuitemradio"
+                >
+                  {t.label}
+                </MenuRow>
+              ))}
+              <div aria-hidden className="my-1 h-px bg-edge/40" />
+            </>
+          )}
+          <MenuRow
+            selected={false}
+            onClick={() => {
+              close();
+              onOpenAccount();
+            }}
+          >
+            Yahoo account &amp; leagues…
+          </MenuRow>
+        </>
+      )}
+    </GearMenu>
   );
 }
